refactor: separar dados em modulos dedicados
