+++ src/data/index.ts
@@ -1,0 +1,4 @@
+export * from "./materiais";
+export * from "./trelicas";
+export * from "./tabelaK";
+export * from "./aco";
