+++ src/data/tabelaK.ts
@@ -1,0 +1,60 @@
+/**
+ * Tabela K segundo a norma, utilizada para obtenção de parâmetros de cálculo.
+ */
+
+// [Kc, C20, C25, C30, C35, C40, C45, C50, ks_CA25, ks_CA50, ks_CA60]
+export type LinhaTabelaK = [
+  number, number, number, number, number, number, number, number,
+  number | null, number | null, number | null
+];
+
+// Índices das colunas na tabela
+export const K_COL_KC = 0;
+export const K_COL_KS = 1;
+export const K_COL_BX = 8;
+
+/**
+ * Valores da Tabela K.
+ */
+export const TABELA_K: LinhaTabelaK[] = [
+  [0.02, 51.9, 41.5, 34.6, 29.6, 25.9, 23.1, 20.8, 0.046, 0.023, 0.019],
+  [0.04, 26.2, 20.9, 17.4, 14.9, 13.1, 11.6, 10.5, 0.047, 0.023, 0.019],
+  [0.06, 17.6, 14.1, 11.7, 10.0, 8.8, 7.8, 7.0, 0.047, 0.024, 0.019],
+  [0.08, 13.3, 10.6, 8.9, 7.6, 6.6, 5.9, 5.3, 0.048, 0.024, 0.020],
+  [0.10, 10.7, 8.6, 7.1, 6.1, 5.4, 4.8, 4.3, 0.048, 0.024, 0.020],
+  [0.12, 9.0, 7.2, 6.0, 5.1, 4.5, 4.0, 3.6, 0.048, 0.024, 0.020],
+  [0.14, 7.8, 6.2, 5.2, 4.5, 3.9, 3.5, 3.1, 0.049, 0.024, 0.020],
+  [0.16, 6.9, 5.5, 4.6, 3.9, 3.4, 3.1, 2.7, 0.049, 0.025, 0.020],
+  [0.18, 6.2, 4.9, 4.1, 3.5, 3.1, 2.7, 2.5, 0.050, 0.025, 0.021],
+  [0.20, 5.6, 4.5, 3.7, 3.2, 2.8, 2.5, 2.2, 0.050, 0.025, 0.021],
+  [0.22, 5.1, 4.1, 3.4, 2.9, 2.6, 2.3, 2.1, 0.051, 0.025, 0.021],
+  [0.24, 4.7, 3.8, 3.2, 2.7, 2.4, 2.1, 1.9, 0.051, 0.025, 0.021],
+  [0.259, 4.4, 3.6, 3.0, 2.5, 2.2, 2.0, 1.8, 0.051, 0.026, 0.021],
+  [0.28, 4.1, 3.3, 2.8, 2.4, 2.1, 1.8, 1.7, 0.052, 0.026, 0.022],
+  [0.30, 3.9, 3.1, 2.6, 2.2, 1.9, 1.7, 1.6, 0.052, 0.026, 0.022],
+  [0.32, 3.7, 3.0, 2.5, 2.1, 1.8, 1.6, 1.5, 0.053, 0.026, 0.022],
+  [0.34, 3.5, 2.8, 2.3, 2.0, 1.8, 1.6, 1.4, 0.053, 0.027, 0.022],
+  [0.36, 3.3, 2.7, 2.2, 1.9, 1.7, 1.5, 1.3, 0.054, 0.027, 0.022],
+  [0.38, 3.2, 2.6, 2.1, 1.8, 1.6, 1.4, 1.3, 0.054, 0.027, 0.023],
+  [0.40, 3.1, 2.5, 2.0, 1.8, 1.5, 1.4, 1.2, 0.055, 0.028, 0.023],
+  [0.42, 2.9, 2.4, 2.0, 1.7, 1.5, 1.3, 1.2, 0.055, 0.028, 0.023],
+  [0.44, 2.8, 2.3, 1.9, 1.6, 1.4, 1.3, 1.1, 0.056, 0.028, 0.023],
+  [0.46, 2.7, 2.2, 1.8, 1.6, 1.4, 1.2, 1.1, 0.056, 0.028, 0.023],
+  [0.48, 2.7, 2.1, 1.8, 1.5, 1.3, 1.2, 1.1, 0.057, 0.028, 0.024],
+  [0.50, 2.6, 2.1, 1.7, 1.5, 1.3, 1.1, 1.0, 0.058, 0.029, 0.024],
+  [0.52, 2.5, 2.0, 1.7, 1.4, 1.2, 1.1, 1.0, 0.058, 0.029, 0.024],
+  [0.54, 2.4, 2.0, 1.6, 1.4, 1.2, 1.1, 1.0, 0.059, 0.029, 0.024],
+  [0.56, 2.4, 1.9, 1.6, 1.4, 1.2, 1.1, 1.0, 0.059, 0.030, 0.025],
+  [0.585, 2.3, 1.8, 1.5, 1.3, 1.2, 1.0, 0.9, 0.060, 0.030, 0.025],
+  [0.60, 2.3, 1.8, 1.5, 1.3, 1.1, 1.0, 0.9, 0.061, 0.030, null],
+  [0.628, 2.2, 1.8, 1.5, 1.3, 1.1, 1.0, 0.9, 0.062, 0.031, null],
+  [0.64, 2.2, 1.7, 1.4, 1.2, 1.1, 1.0, 0.9, 0.062, null, null],
+  [0.66, 2.1, 1.7, 1.4, 1.2, 1.1, 0.9, 0.8, 0.063, null, null],
+  [0.68, 2.1, 1.7, 1.4, 1.2, 1.1, 0.9, 0.8, 0.063, null, null],
+  [0.70, 2.0, 1.6, 1.4, 1.2, 1.0, 0.9, 0.8, 0.064, null, null],
+  [0.72, 2.0, 1.6, 1.3, 1.2, 1.0, 0.9, 0.8, 0.065, null, null],
+  [0.74, 2.0, 1.6, 1.3, 1.1, 1.0, 0.9, 0.8, 0.065, null, null],
+  [0.76, 2.0, 1.6, 1.3, 1.1, 1.0, 0.9, 0.8, 0.066, null, null],
+  [0.772, 1.9, 1.5, 1.3, 1.1, 1.0, 0.9, 0.8, 0.067, null, null],
+];
+
